Config again firebase service, init once.
