+++ src/main/java/gigjob/firebase/InitFirebaseApp.java
@@ -1,0 +1,34 @@
+package gigjob.firebase;
+
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseOptions;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.event.EventListener;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.stereotype.Component;
+
+@Component
+public class InitFirebaseApp {
+    @Value("${firebase.google-credentials}")
+    private String googleCredential;
+    @Value("${firebase.bucket-name}")
+    private String bucketName;
+
+    @EventListener
+    public void initialize(ApplicationReadyEvent event) {
+        try {
+            GoogleCredentials googleCredentials = GoogleCredentials
+                    .fromStream(new ClassPathResource(googleCredential).getInputStream());
+            FirebaseOptions options = FirebaseOptions.builder()
+                    .setCredentials(googleCredentials)
+                    .setStorageBucket(bucketName)
+                    .build();
+
+            FirebaseApp.initializeApp(options);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}
