+++ src/main/java/gigjob/firebase/authentication/UserManagementService.java
@@ -1,0 +1,14 @@
+package gigjob.firebase.authentication;
+
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseAuthException;
+import com.google.firebase.auth.UserRecord;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserManagementService {
+
+    public UserRecord getFirebaseUserById(String uid) throws FirebaseAuthException {
+        return FirebaseAuth.getInstance().getUser(uid);
+    }
+}
