--- src/main/java/gigjob/firebase/messaging/FirebaseMessagingConfig.java
@@ -1,33 +1,0 @@
-package gigjob.firebase.messaging;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import com.google.firebase.messaging.FirebaseMessaging;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-
-import java.io.IOException;
-
-@Configuration
-@RequiredArgsConstructor
-public class FirebaseMessagingConfig {
-    @Value("${firebase.google-credentials}")
-    private String googleCredential;
-
-    @Bean
-    public FirebaseMessaging firebaseMessaging() throws IOException {
-        GoogleCredentials googleCredentials = GoogleCredentials
-                .fromStream(new ClassPathResource(googleCredential).getInputStream());
-        FirebaseOptions firebaseOptions = FirebaseOptions
-                .builder()
-                .setCredentials(googleCredentials)
-                .build();
-        FirebaseApp app = FirebaseApp.initializeApp(firebaseOptions, "GIGJOB");
-        return FirebaseMessaging.getInstance(app);
-    }
-
-}
